+++ agrest-cayenne/src/test/java/io/agrest/cayenne/encoder/EncoderServiceTest.java
@@ -1,5 +1,6 @@
 package io.agrest.cayenne.encoder;
 
+import io.agrest.DataResponse;
 import io.agrest.ResourceEntity;
 import io.agrest.RootResourceEntity;
 import io.agrest.SimpleObjectId;
@@ -127,7 +128,7 @@ public class EncoderServiceTest extends CayenneNoDbTest {
 
     private String toJson(Object object, ResourceEntity<?> resourceEntity) {
         Encoder encoder = encoderService.dataEncoder(resourceEntity);
-        return toJson(encoder, Collections.singletonList(object));
+        return toJson(encoder, DataResponse.of(200, Collections.singletonList(object)));
     }
 
     private String toJson(Encoder encoder, Object value) {