+++ agrest-engine/src/main/java/io/agrest/property/ToManyEntityResultReader.java
@@ -22,7 +22,7 @@ public class ToManyEntityResultReader extends NestedEntityResultReader<ToManyRes
     @Override
     public List<?> value(Object root) {
         AgObjectId id = readId(root);
-        List<?> value = entity.getResult(id);
+        List<?> value = entity.getDataWindow(id);
         return value != null ? value : Collections.emptyList();
     }
 }