+++ agrest-engine/src/main/java/io/agrest/runtime/processor/select/SelectContext.java
@@ -48,8 +48,10 @@ public class SelectContext<T> extends BaseProcessingContext<T> {
      * @since 1.24
      */
     public DataResponse<T> createDataResponse() {
-        List<? extends T> objects = this.entity != null ? this.entity.getResult() : Collections.emptyList();
-        return DataResponse.of(getStatus(), objects, encoder);
+        // support null ResourceEntity for cases with custom terminal stages
+        return entity != null
+                ? DataResponse.of(getStatus(), entity.getDataWindow(), entity.getResult().size(), encoder)
+                : DataResponse.of(getStatus());
     }
 
     public boolean isById() {