+++ agrest-engine/src/test/java/io/agrest/DataResponseTest.java
@@ -17,10 +17,11 @@ public class DataResponseTest {
         List<Tr> objects = asList(new Tr(), new Tr());
         Encoder encoder = mock(Encoder.class);
 
-        DataResponse<Tr> response = DataResponse.of(201, objects, encoder);
+        DataResponse<Tr> response = DataResponse.of(201, objects, 2, encoder);
         assertNotNull(response);
         assertEquals(201, response.getStatus());
-        assertSame(objects, response.getObjects());
+        assertSame(objects, response.getData());
+        assertEquals(2, response.getTotal());
         assertSame(encoder, response.getEncoder());
     }